backend/django/requirements.txt
Django>=5.0
psycopg[binary]
python-dotenv==1.2.1
djangorestframework==3.16.1